moved database to another file
