+++ app/domain/database.go
@@ -1,0 +1,15 @@
+package domain
+
+type DatabaseConfig struct {
+	Driver   string
+	Host     string
+	Database string
+	Port     string
+	Username string
+	Password string
+}
+
+type Database interface {
+	Init()
+	Clear()
+}
